chore: upgrade WASI to @0.2.9
- Updated all WIT deps from unified WebAssembly/WASI repo (individual repos archived at 0.2.8)
- Converted ts-runtime-mcp from binary to cdylib (main.rs -> lib.rs)
- Fixed terminal-info u32->u16 type mismatch in web-agent-tui backend.rs
- Fixed ErrorCode variant format in wasi-http-impl FutureIncomingResponse
- Added terminal-info custom WIT interface
- Updated transpile.mjs and build.yml for ts_runtime_mcp.wasm (underscore)
- Removed per-crate WIT deps (now using shared runtime/wit/deps/)
- All E2E tests passing (wasm-runtime: 34, tui: 11, jspi-context: 3)

+++ runtime/crates/web-agent-tui/src/backend.rs
@@ -136,7 +136,7 @@ impl<W: Write> Backend for WasiBackend<W> {
         // Poll terminal size from JS via WIT import
         use crate::bindings::terminal::info::size::get_terminal_size;
         let dims = get_terminal_size();
-        Ok(Size::new(dims.cols, dims.rows))
+        Ok(Size::new(dims.cols as u16, dims.rows as u16))
     }
 
     fn flush(&mut self) -> IOResult<()> {
